resource_manage/requirements.txt
Django==3.2.14
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
django-cors-headers==3.13.0
numpy==1.21.6
pandas==1.3.5
scipy==1.7.3
drf-yasg==1.21.3
django-filter==22.1
mysqlclient==2.1.1
gunicorn==20.1.0
httptools==0.4.0
uvloop==0.16.0
protobuf==3.20.1
macropodus==0.0.7
django-redis==5.2.0
uvicorn==0.18.2
ChineseTone==0.1.4

